+++ userService/src/main/java/com/Golisoda/userService/Service/userService.java
@@ -1,6 +1,8 @@
 package com.Golisoda.userService.Service;
 
 import com.Golisoda.userService.Dto.userDto;
+import com.Golisoda.userService.Exception.userException;
+import com.Golisoda.userService.Util.AddressMapper;
 import com.Golisoda.userService.Util.UserMapper;
 import com.Golisoda.userService.dao.userDao;
 import org.slf4j.Logger;
@@ -20,32 +22,56 @@ public class userService {
     userDao userdao;
     @Autowired
     UserMapper usermapper;
+    @Autowired
+    addressService addressService;
+    @Autowired
+    AddressMapper addressmapper;
     public userDto createUSer(userDto user) {
 
         log.info("Adding new user {}",user);
         var userForDB = usermapper.toUser(user);
-    userForDB.setUser_id(UUID.randomUUID().toString());
+        var userID = UUID.randomUUID().toString();
+    userForDB.setUser_id(userID);
+    var address = user.getAddress();
+    var addressForDB = addressmapper.toAddress(address);
+    addressForDB.setUserId(userID);
+    var addressID = UUID.randomUUID().toString();
+    addressForDB.setAddress_id(addressID);
+    addressForDB.setUser_type("user");
+    addressService.addAddress(addressForDB);
+    userForDB.setAddress(addressForDB);
     return usermapper.toUserDto(userdao.save(userForDB));
 
     }
 
-    public userDto readUser(String  id) {
-        log.info("Fetching data from DB with user id {}",id);
-        var userFromDB= userdao.findById(id)
-                .orElseThrow(()-> new IllegalArgumentException("User not found"));
+    public userDto readUser(String  email_id) {
+        log.info("Fetching data from DB with user id {}",email_id);
+        var userFromDB= userdao.findByEmailid(email_id);
+        if(userFromDB==null){
+            throw new userException("User with email id %s is not available".formatted(email_id));
+        }
         return usermapper.toUserDto(userFromDB);
     }
 
-    public String deleteUser(String id) {
-        log.info("Deleting user with id {} from DB",id);
+    public String deleteUser(String email_id) {
+        var userDeletedDB= userdao.findByEmailid(email_id);
+        if(userDeletedDB==null){
+            throw new userException("User with email id %s is not available".formatted(email_id));
+        }
+        var id = userDeletedDB.getUser_id();
+        log.info("Deleting user with email id {} from DB",email_id);
         userdao.deleteById(id);
-        return "Deleted user with id" + id;
+        return "Deleted user with email id" + email_id;
     }
 
-    public userDto updateUser(String id, userDto ud) {
-        log.info("Updating user data in DB with user id {}",id);
-        var existingud = userdao.findById(id)
-                .orElseThrow(()-> new IllegalArgumentException("User not found"));
+    public userDto updateUser(String email_id, userDto ud) {
+        log.info("Updating user data in DB with email id {}",email_id);
+        var existinguserDB = userdao.findByEmailid(email_id);
+        if(existinguserDB==null){
+            throw new userException("User with email id %s is not available".formatted(email_id));
+        }
+
+        var existingud = usermapper.toUserDto(existinguserDB);
         existingud.setF_name(ud.getF_name());
         existingud.setL_name(ud.getL_name());
         existingud.setDob(ud.getDob());
@@ -57,9 +83,11 @@ public class userService {
         existingud.setAddress(ud.getAddress());
         existingud.setPhone_no(ud.getPhone_no());
         existingud.setAlt_phone_no(ud.getAlt_phone_no());
-        existingud.setEmail_id(ud.getEmail_id());
+        existingud.setEmailid(ud.getEmailid());
 
-        return usermapper.toUserDto(userdao.save(existingud));
+        var resultUser = usermapper.toUser(existingud);
+
+        return usermapper.toUserDto(userdao.save(resultUser));
